// src/schema/auth.schema.ts
import { z } from "zod";
const loginSchema = z.object({
  email: z.string().min(1, "Email is required").email("Enter a valid email address"),
  password: z.string().min(1, "Password is required").min(8, "Password must be at least 8 characters"),
  remember: z.boolean().optional(),
});

export { loginSchema };